Finally got it. After making it work for positive and negative speeds (had to back out some earlier shortcuts in gradient computation that lead to all gradients always lying in the first quadrant).
And there is a big issue with the idea of using direction-dependent
speed for sailboat path planning: what counts for a boat is NOT the
direction of the wavefront, but JUST the wind direction (and
currents). What's happening now is that the wavefront aligns itself
with certain directions that depend on the speed profile, and stick to
it, creating straight edges in the crossing time map. Probably around
extrema in the speed map. Or, what I also observed when there's a dead
angle, the wavefront won't go in there at all, it somehow gets stuck
in a straight line whose normal lies in the dead sector of the speed
map. Given that we compute the angle relative to the wavefront
direction, of course it then does not propagate anymore, because it
thinks it cannot.

So, this needs more thought, the principle is flawed. The gradient
direction of the embedding function phi still matters, because it
encodes from where we arrived at a certain point, but the outgoing
velocity should not directly depend on it. For example, make some
estimate of how far we could get by turning the boat up or downwind,
wihin some time, given that we are now at some some heading along the
propagation direction.